chore: capture google maps screenshot using html2canvas
