+++ src/app.js
@@ -6,6 +6,12 @@ import {
 window.mymap = null
 
 const start = () => {
+  /* eslint-disable no-unused-vars */
+  const centerCity = {
+    lat: 47.6061389,
+    lng: -122.3328481
+  }
+
   const center = {
     lat: 47.242384,
     lng: -122.463904
@@ -28,11 +34,17 @@ const start = () => {
 
   // Buttons
   const buttonBasic = document.getElementById('screenshot-basic')
+  const buttonBasicCanvas = document.getElementById('screenshot-basic-canvas')
   const buttonLeaflet = document.getElementById('screenshot-leaflet')
 
   buttonBasic.addEventListener('click', (e) => {
     console.log(map)
     map.screenshot()
+  })
+
+  buttonBasicCanvas.addEventListener('click', (e) => {
+    console.log(map)
+    map.screenshotCanvas()
   })
 
   buttonLeaflet.addEventListener('click', (e) => {
