+++ src/index.html
@@ -48,6 +48,10 @@
         Screenshot (Static MAPS API)
       </button>
 
+      <button id="screenshot-basic-canvas">
+        Screenshot (html2canvas)
+      </button>
+
       <p>
         This demo renders a Google Map using the Google Maps API.
       </p>
