+++ src/map-google/mutant.js
@@ -1,6 +1,7 @@
 import Webmap from '../map/map'
 import GoogleMap from './basic'
 import { leafletDrawOptions } from '../map-draw/constants'
+import { screenshotCanvas } from './utils'
 
 /**
  * Sub class for rendering a Google Map insidea a Leaflet web map using the LeafletJS GoogleMutant plugin.
@@ -53,7 +54,6 @@ class GoogleMapLeaflet extends Webmap {
     this.drawControl = new L.Control.Draw(leafletDrawOptions)
 
     this.map.addControl(this.drawControl, {
-      mapTypeId: 'satellite',
       center: {
         lat: process.env.MAP_LAT,
         lng: process.env.MAP_LON
@@ -100,22 +100,7 @@ class GoogleMapLeaflet extends Webmap {
    * Uses the html2canvas library to capture screenshot in a canvas.
    */
   screenshot () {
-    const mapContainer = document.getElementById(this.mapId)
-
-    html2canvas(mapContainer, { useCORS: true }).then((canvas) => {
-      console.log(canvas)
-      canvas.toBlob((blob) => {
-        const fileURL = URL.createObjectURL(blob)
-        const link = document.createElement('a')
-
-        link.href = fileURL
-        link.setAttribute('download', 'file.png')
-        document.body.appendChild(link)
-
-        link.click()
-        document.body.removeChild(link)
-      })
-    })
+    screenshotCanvas(this.mapId)
   }
 }
